feat 페이저블 기능 추가 - 페이징 아주 살짝 해놓은거라 손 좀 봐야됨

+++ src/main/java/com/bs/search/domain/BookEntity.java
@@ -43,32 +43,16 @@ public class BookEntity {
     private String publisher;
 
     @Column(name = "price")
-    private int price;
+    private long price;
 
     @Column(name = "salePrice")
-    private int salePrice;
+    private long salePrice;
 
     @Column(name = "thumbnail" , length = 1000)
     private String thumbnail;
 
     @Column(name = "status")
     private String status;
-
-//  @OneToMany
-//    @JoinTable(name = "authors_book",
-//            joinColumns = @JoinColumn(name = "book_id"),
-//            inverseJoinColumns = @JoinColumn(name = "author_id"))
-//    private List<AuthorsEntity> authorsList =  new ArrayList<>();
-//
-//    @OneToMany
-//    @JoinTable(name = "translators_book",
-//            joinColumns = @JoinColumn(name = "book_id"),
-//            inverseJoinColumns = @JoinColumn(name = "author_id"))
-//    private List<TranslatorsEntity> translatorsList;
-//    @OneToMany(cascade =CascadeType.ALL, mappedBy = "book", fetch = FetchType.LAZY)
-//    private List<AuthorsEntity> authorsList =  new ArrayList<>();
-
-
 
     @OneToMany( fetch = FetchType.LAZY)
     @JoinColumn(name="author_id")
@@ -78,6 +62,4 @@ public class BookEntity {
     @JoinColumn(name="trans_id")
     private Set<TranslatorsEntity> translatorAuthorEntitys;
 
-//    @OneToMany(mappedBy = "book")
-//    private List<TranslatorsEntity> translatorsList;
 }
